+++ DrawingApp/DrawingApp/Utility/Extensions/CGTypes+Extensions.swift
@@ -1,0 +1,13 @@
+//
+//  CGTypes+Extensions.swift
+//  DrawingApp
+//
+//  Created by 송태환 on 2022/03/02.
+//
+
+import UIKit
+
+extension CGSize: SizeBuilder {}
+extension CGPoint: PointBuilder {}
+extension CGFloat: AlphaBuilder {}
+extension CGRect: RectangleBuilder {}
