+++ DrawingApp/DrawingApp/Utility/Extensions/String+Alphanumeric.swift
@@ -1,0 +1,12 @@
+//
+//  String+Alphanumeric.swift
+//  DrawingApp
+//
+//  Created by 송태환 on 2022/03/07.
+//
+
+import Foundation
+
+extension String {
+    static let alphanumeric = "abcdefghjklmnpqrstuvwxyz12345789"
+}
